detect folder before rm

+++ replayer/lib/handler/multiple.ts
@@ -36,7 +36,9 @@ const createTemporaryFolders = (
 	const resultParamsTempFolder = path.join(workspace, '.result-params-temp');
 	if (!env.isOnChildProcess()) {
 		// not in child process, delete the result temp folder
-		fs.rmdirSync(resultParamsTempFolder, { recursive: true });
+		if (fs.existsSync(resultParamsTempFolder)) {
+			fs.rmdirSync(resultParamsTempFolder, { recursive: true });
+		}
 	}
 	if (!fs.existsSync(resultParamsTempFolder)) {
 		fs.mkdirSync(resultParamsTempFolder);
